% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{addDimorphism}
\alias{addDimorphism}
\title{Adds two columns with data from Chromatinsight (loaded using getRegionData)
to a table containing connected regions from Hi-C experiments.}
\usage{
addDimorphism(
  data,
  dimorphismData,
  suffix = "",
  overwrite = FALSE,
  counterstep = 0,
  sourceColumn = "disparity"
)
}
\description{
Adds two columns with data from Chromatinsight (loaded using getRegionData)
to a table containing connected regions from Hi-C experiments.
}
\examples{
dimorphismMono = getRegionData(folderPath, prefix = "mono", totRandomStates = 5)
HiCDataIncludingDimorphism = addDimorphism(data = HiCData, dimorphismData = dimorphismMono)
}
