% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{addFDR}
\alias{addFDR}
\title{Generates a new table with an extra column displaying the FDR of the
disparity.
The input is two output tables from getRegionData:
tableObserved: the table with the data including the result from the
Chromatinsight analysis.
tableRnd: the table with the same data, but randomising the category
labels (in testPrediction, randomize = True).}
\usage{
addFDR(regionDataObserved, regionDataRnd, verbose = TRUE)
}
\description{
Generates a new table with an extra column displaying the FDR of the
disparity.
The input is two output tables from getRegionData:
tableObserved: the table with the data including the result from the
Chromatinsight analysis.
tableRnd: the table with the same data, but randomising the category
labels (in testPrediction, randomize = True).
}
\examples{
newTable = addFDR(tableObserved, tableRnd)
}
