% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatinsight.tools.R
\docType{package}
\name{chromatinsight.tools}
\alias{chromatinsight.tools}
\title{chromatinsight.tools v1.3: A set of methods for R
used in tandem with Chromatinsight}
\description{
Author: Marco Trevisan-Herraz, PhD
Computational Epigenomics Laboratory
Newcastle University, UK
2019-2021
}
\section{Functions}{

pileup
drawpile
drawpilegb
densityPile
getRegionData
fillTableWithGeneTypes
}

