% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compareDistributions}
\alias{compareDistributions}
\title{Displays the violin plots of the distributions in two analyses
retrieved using getRegionData. Especially useful to compare random vs
observed data.}
\usage{
compareDistributions(
  regionData1,
  regionData2,
  label1 = "first",
  label2 = "second",
  yLabel = "data",
  graphTitle = "comparison",
  filename = "",
  filedim = c(1200, 1200)
)
}
\description{
Displays the violin plots of the distributions in two analyses
retrieved using getRegionData. Especially useful to compare random vs
observed data.
}
