% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{countGeneTypes}
\alias{countGeneTypes}
\title{Counts the gene types for each of the
gene types dowloaded from Biomart.}
\usage{
countGeneTypes(myGeneTypes, SDGeneTypes)
}
\description{
Counts the gene types for each of the
gene types dowloaded from Biomart.
}
