% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{densitypile}
\alias{densitypile}
\title{You give it the output of two pileup function
and it shows the graph comparing them.}
\usage{
densitypile(
  datafem,
  datamal,
  start = 0,
  plus = 1000,
  highstart = 0,
  highplus = 0,
  sex = "both",
  histmod = "ac",
  chrom = "",
  opacity = 0.5,
  label = ""
)
}
\description{
You give it the output of two pileup function
and it shows the graph comparing them.
}
