% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fillTableWithGeneTypes}
\alias{fillTableWithGeneTypes}
\title{Merges the information from Biomart and the data from Chromatinsight,
retrieved using getRegionData.}
\usage{
fillTableWithGeneTypes(data, BioMartChr, SDGeneTypes)
}
\description{
Merges the information from Biomart and the data from Chromatinsight,
retrieved using getRegionData.
}
