% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getheatmap}
\alias{getheatmap}
\title{Calculates the heatmap of a set of ChromHMM output binary files.
Parameters:
pattern: to select the files from the filename. Asterisks (*) as wildcard.
coord: the coordinates of the region to extract, such as
chrX:73,040,484-73,072,558 (this parameter is mandatory)
histmod: the name of the column from which extract the data, such as the
histone mark. Default is "H3K27ac".
windowsize: the number of bases in each ChromHMM window. Default is 200.
direc: directory or folder where the data are (omit to use the working
directory).
filename: the filename to save a 1800x1000 px image with the heatmap. If
omitted, it will just show it in the plot area.}
\usage{
getheatmap(
  pattern = "*",
  coord = "",
  histmod = "H3K27ac",
  windowsize = 200,
  filename = "",
  direc = ""
)
}
\description{
Calculates the heatmap of a set of ChromHMM output binary files.
Parameters:
pattern: to select the files from the filename. Asterisks (*) as wildcard.
coord: the coordinates of the region to extract, such as
chrX:73,040,484-73,072,558 (this parameter is mandatory)
histmod: the name of the column from which extract the data, such as the
histone mark. Default is "H3K27ac".
windowsize: the number of bases in each ChromHMM window. Default is 200.
direc: directory or folder where the data are (omit to use the working
directory).
filename: the filename to save a 1800x1000 px image with the heatmap. If
omitted, it will just show it in the plot area.
}
\examples{
getheatmap(pattern = "mono*",
           coord = "chrX:130,822,807-130,964,890",
           histmod = "H3K4me1",
           filename = "C:/data/Firre_reduced_me.png")
}
