% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getHistModLevels2}
\alias{getHistModLevels2}
\title{Deprecated version of the function getHistModLevels that was using the file
name rather than the grouping file.
Adds five columns to the region table, including additional ChIP-seq data
for all windows and samples corresponding to a genomic region:
\enumerate{
\item Ratio of windows that have the histone modification.
\item Average for sample group A.
\item Average for sample group B.
\item Relative ratio for sample group A.
\item Relative ratio for sample group B.
}}
\usage{
getHistModLevels2(
  data,
  direc = "",
  histMod = "ac",
  prefixA = "",
  nameA = "1st",
  prefixB = "",
  nameB = "2nd",
  suffix = "real",
  windowSize = 200,
  verbose = TRUE
)
}
\description{
Deprecated version of the function getHistModLevels that was using the file
name rather than the grouping file.
Adds five columns to the region table, including additional ChIP-seq data
for all windows and samples corresponding to a genomic region:
\enumerate{
\item Ratio of windows that have the histone modification.
\item Average for sample group A.
\item Average for sample group B.
\item Relative ratio for sample group A.
\item Relative ratio for sample group B.
}
}
\examples{
getHistModLevels2(data,
                  direc = "",
                  histMod = "ac",
                  prefixA = "mono*S*fem*NCMLS*",
                  nameA = "fem",
                  prefixB = "mono*S*mal*NCMLS*",
                  nameB = "mal",
                  suffix = "real",
                  windowSize = 200,
                  verbose = TRUE)
}
