% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getmatrix}
\alias{getmatrix}
\title{Extracts the matrix of histone modifications to be used to
represent the heatmap.}
\usage{
getmatrix(
  direc = "",
  pattern = "*",
  coord = "",
  start = 1,
  end = -1,
  histmod = "H3K27ac",
  windowsize = 200,
  chromosome = ""
)
}
\description{
Extracts the matrix of histone modifications to be used to
represent the heatmap.
}
