% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getmatrix2}
\alias{getmatrix2}
\title{Deprecated version of getmatrix that uses the pattern of the filename
rather than the grouping file.
Extracts the matrix of histone modifications to be used to
represent the heatmap.}
\usage{
getmatrix2(
  direc = "",
  pattern = "*",
  coord = "",
  start = 1,
  end = -1,
  histmod = "H3K27ac",
  windowsize = 200,
  chromosome = ""
)
}
\description{
Deprecated version of getmatrix that uses the pattern of the filename
rather than the grouping file.
Extracts the matrix of histone modifications to be used to
represent the heatmap.
}
