% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getRegionData}
\alias{getRegionData}
\title{Loads the output text files of Chromatinsight (currently Python only),
Using as output the median of the number of trials porformed (in the
totRandomStates variable).}
\usage{
getRegionData(
  myPath = "",
  histmod = "ac",
  prefix = "prefix",
  suffix = "",
  totRandomStates = 11
)
}
\description{
Loads the output text files of Chromatinsight (currently Python only),
Using as output the median of the number of trials porformed (in the
totRandomStates variable).
}
