% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pileup}
\alias{pileup}
\title{You give it a grouping file and key to find a set files of ChromHMM binaries
and it adds up the "ones" and "zeroes" that are at the same position for the
set of files, dividing by the number of files.
So it gives the ratio of samples having a "one" at a specific position
ie, the ratio of samples being H3K27ac or H3K4me1 at a ChromHMM bin.
If the parameter "key" is omitted, then all the files will be used.}
\usage{
pileup(grouping = "", key = "", direc = "", chrom = "")
}
\description{
You give it a grouping file and key to find a set files of ChromHMM binaries
and it adds up the "ones" and "zeroes" that are at the same position for the
set of files, dividing by the number of files.
So it gives the ratio of samples having a "one" at a specific position
ie, the ratio of samples being H3K27ac or H3K4me1 at a ChromHMM bin.
If the parameter "key" is omitted, then all the files will be used.
}
