% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{splitCoords}
\alias{splitCoords}
\title{To split the genomic coordinates}
\usage{
splitCoords(coords = "")
}
\description{
To split the genomic coordinates
}
\examples{
splitCoords(coords = "chrX:10,000-25,000")
is the same as
list(chrom = "chrX", start = "10000", end = "25000")
}
