% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pileup}
\alias{pileup}
\title{You give it the filename prefix to find a set files of ChromHMM binaries
and it adds up the "ones" and "zeroes" that are at the same position for the
set of files, dividing by the number of files.
So it gives the ratio of samples having a "one" at a specific position
ie, the ratio of samples being H3K27ac or H3K4me1 at a ChromHMM bin.}
\usage{
pileup(prefix = "", direc = "", chrom = "")
}
\description{
You give it the filename prefix to find a set files of ChromHMM binaries
and it adds up the "ones" and "zeroes" that are at the same position for the
set of files, dividing by the number of files.
So it gives the ratio of samples having a "one" at a specific position
ie, the ratio of samples being H3K27ac or H3K4me1 at a ChromHMM bin.
}
\examples{
pileup(prefix = "mono*S*mal", direc = "/data/", chrom = "chr10")
Will merge data from files (within /data folder)
monocytes_S2901_mal_DATA_binary.txt
monocytes_S3454_mal_DATA_binary.txt
monocytes_S5715_mal_DATA_binary.txt
etc
}
